Refactor imports and update CLI entry
